COMP: fix warnings of unused vars, signed/unsigned comparisons, and bool/int conversions
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=6434
git-svn-id: http://svn.slicer.org/Slicer4/trunk@6434 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/vtkTeem/vtkNRRDWriter.cxx
@@ -101,8 +101,8 @@ void vtkNRRDWriter::vtkImageDataInfoToNrrdInfo(vtkImageData *in, int &nrrdKind, 
         nrrdKind = nrrdKindRGBAColor;
         break;
       default:
-        int numGrad = this->DiffusionGradients->GetNumberOfTuples();
-        int numBValues = this->BValues->GetNumberOfTuples();
+        size_t numGrad = this->DiffusionGradients->GetNumberOfTuples();
+        size_t numBValues = this->BValues->GetNumberOfTuples();
         if (numGrad == numBValues && numGrad == numComp && numGrad>6)
           {
           nrrdKind = nrrdKindList;
